+++ src/Atc.Cosmos.EventStore.Cqrs/Testing/CommandHandlerExtensions.cs
@@ -7,5 +7,12 @@ public static class CommandHandlerExtensions
         params object[] events)
         where TCommand : ICommand
         => new CommandHandlerTester<TCommand>(handler)
-            .GivenEvents(events);
+            .GivenEvents(() => Task.FromResult(events));
+
+    public static ICommandWhen<TCommand> GivenStreamContainingEvents<TCommand>(
+        this ICommandHandler<TCommand> handler,
+        Func<Task<object[]>> eventsFactory)
+        where TCommand : ICommand
+        => new CommandHandlerTester<TCommand>(handler)
+            .GivenEvents(eventsFactory);
 }