+++ src/Atc.Cosmos.EventStore.Cqrs/Testing/ICommandGiven.cs
@@ -3,5 +3,6 @@ namespace Atc.Cosmos.EventStore.Cqrs.Testing;
 public interface ICommandGiven<in TCommand>
     where TCommand : ICommand
 {
-    ICommandWhen<TCommand> GivenEvents(params object[] events);
+    ICommandWhen<TCommand> GivenEvents(
+        Func<Task<object[]>> eventsFactory);
 }